requirements-dev.txt
nose==1.3.4
flake8==2.3.0
pep8==1.6.1
rednose>=0.4.3
coverage==3.7.1
PyYAML>=3.11
mock>=1.3.0
-e git://github.com/cogniteev/celery.git@v3.1.17-cogniteev-1#egg=celery
